+++ gulp/tasks/jspm.js
@@ -20,6 +20,28 @@ Object.keys(config.builder.bundles).forEach(b => {
   });
 });
 
+gulp.task(`jspm-dev-bundle`, () => {
+  const builder = new jspm.Builder('./');
+  builder.config({
+    buildCSS: true,
+    buildHTML: true,
+    // rootURL: '.',
+    separateCSS: true,
+    paths: {
+      'github:*': 'jspm_packages/github/*',
+      'npm:*': 'jspm_packages/npm/*',
+      'components/*': 'app/components/*',
+      'common/*': 'app/common/*',
+      'bundles/*': 'app/bundles/*'
+    }
+  });
+  return builder.bundle(
+    `${paths.base}/components/boot.js + universe + d3-svg-legend + d3-tip + URIjs + angular-ui-grid - [${paths.base}/**/*] - [${paths.base}/**/*!css] - [${paths.base}/**/*!text] - [${paths.base}/**/*!md]`,
+    `${paths.temp}/${paths.bundles}/deps-bundle.js`,
+    config.builder.bundle
+  );
+});
+
 // copy bundles to dist folder
 gulp.task('jspm-copy-bundles', () => {
   return gulp.src([`${paths.temp}/${paths.bundles}/*.{js,map,html}`], {base: paths.temp})
@@ -68,6 +90,13 @@ gulp.task('jspm-build-app', cb => {
   runSequence('jspm-app-bundle',
               ['jspm-copy-config', 'jspm-copy-bundles', 'jspm-copy-bundles-css', 'jspm-treemaps'],
               cb);
+});
+
+gulp.task('jspm-dev-build', cb => {
+  runSequence(
+    'jspm-mkdir',
+    'jspm-dev-bundle',
+  cb);
 });
 
 gulp.task('jspm-build', cb => {